# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SoftFactory API",
    "version": "1.0.0",
    "description": "Auto-generated from Flask route decorators."
  },
  "paths": {
...
    "/api/errors/log": {
      "post": {
        "summary": "Log Error",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/errors/patterns": {
      "get": {
        "summary": "Get Error Patterns",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/errors/patterns/{pattern_id}/prevention": {
      "get": {
        "summary": "Get Pattern Prevention",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/errors/patterns/{pattern_id}/resolve": {
      "post": {
        "summary": "Resolve Pattern",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/errors/recent": {
      "get": {
        "summary": "Get Recent Errors",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}